requirements.dev.txt
codecov==2.0.17
factory-boy==2.12.0
ipython==7.13.0
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.8.0
recommonmark==0.6.0
sphinx==2.4.4
sphinx_rtd_theme==0.4.3
sphinx-autobuild==0.7.1
mypy==0.770
django-stubs==1.5.0
whitenoise==5.0.1
werkzeug==1.0.0